requirements.txt
certifi==2020.12.5
chardet==4.0.0
et-xmlfile==1.1.0
future==0.18.2
gemweb @ git+https://github.com/BeeGroup-cimne/beegweb@302be91cd7be8f5a3fd043fb49c2d3705980fff6
happybase==1.2.0
idna==2.10
mrjob==0.7.4
numpy==1.21.0
openpyxl==3.0.7
pandas==1.2.5
ply==3.11
pure-sasl==0.6.2
pycrypto==2.6.1
PyHive==0.6.4
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sasl==0.3.1
six==1.15.0
thrift==0.13.0
thrift-sasl==0.4.3
thriftpy2==0.4.14
urllib3==1.26.4
xlrd==2.0.1
xmltodict==0.12.0
